Moved models under standalone package.
