+++ core-rules/src/main/java/org/hisp/dhis/android/rules/RuleExecution.java
@@ -1,5 +1,7 @@
 package org.hisp.dhis.android.rules;
 
+import org.hisp.dhis.android.rules.models.RuleEffect;
+
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Callable;
