+++ core-rules/src/main/java/org/hisp/dhis/android/rules/models/RuleAttributeValue.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.rules;
+package org.hisp.dhis.android.rules.models;
 
 import com.google.auto.value.AutoValue;
