+++ core-rules/src/main/java/org/hisp/dhis/android/rules/models/RuleVariableAttribute.java
@@ -1,8 +1,6 @@
-package org.hisp.dhis.android.rules;
+package org.hisp.dhis.android.rules.models;
 
 import com.google.auto.value.AutoValue;
-
-import java.util.Map;
 
 import javax.annotation.Nonnull;
 
@@ -20,14 +18,14 @@ public abstract class RuleVariableAttribute extends RuleVariable {
             @Nonnull String trackedEntityAttribute, @Nonnull RuleValueType trackedEntityAttributeType) {
         return new AutoValue_RuleVariableAttribute(name, trackedEntityAttribute, trackedEntityAttributeType);
     }
-
-    @Nonnull
-    RuleVariableValue value(@Nonnull Map<String, RuleAttributeValue> valueMap) {
-        if (valueMap.containsKey(trackedEntityAttribute())) {
-            RuleAttributeValue value = valueMap.get(trackedEntityAttribute());
-            return RuleVariableValue.create(value.value(), trackedEntityAttributeType());
-        }
-
-        return RuleVariableValue.create(trackedEntityAttributeType());
-    }
+//
+//    @Nonnull
+//    RuleVariableValue value(@Nonnull Map<String, RuleAttributeValue> valueMap) {
+//        if (valueMap.containsKey(trackedEntityAttribute())) {
+//            RuleAttributeValue value = valueMap.get(trackedEntityAttribute());
+//            return RuleVariableValue.create(value.value(), trackedEntityAttributeType());
+//        }
+//
+//        return RuleVariableValue.create(trackedEntityAttributeType());
+//    }
 }
