+++ core-rules/src/main/java/org/hisp/dhis/android/rules/models/RuleVariableDataElement.java
@@ -1,0 +1,12 @@
+package org.hisp.dhis.android.rules.models;
+
+import javax.annotation.Nonnull;
+
+abstract class RuleVariableDataElement extends RuleVariable {
+
+    @Nonnull
+    public abstract String dataElement();
+
+    @Nonnull
+    public abstract RuleValueType dataElementType();
+}
