+++ core-rules/src/main/java/org/hisp/dhis/android/rules/models/RuleVariableNewestEvent.java
@@ -1,0 +1,17 @@
+package org.hisp.dhis.android.rules.models;
+
+import com.google.auto.value.AutoValue;
+
+@AutoValue
+public abstract class RuleVariableNewestEvent extends RuleVariableDataElement {
+
+//    @Nonnull
+//    RuleVariableValue value(@Nonnull Map<String, List<RuleDataValue>> valueMap) {
+//        List<RuleDataValue> ruleDataValues = valueMap.get(dataElement());
+//        if (ruleDataValues != null && !ruleDataValues.isEmpty()) {
+//            return RuleVariableValue.create(ruleDataValues.get(0).value(), dataElementType());
+//        }
+//
+//        return RuleVariableValue.create(dataElementType());
+//    }
+}
