+++ core-rules/src/test/java/org/hisp/dhis/android/rules/RuleAttributeValueTests.java
@@ -1,5 +1,6 @@
 package org.hisp.dhis.android.rules;
 
+import org.hisp.dhis.android.rules.models.RuleAttributeValue;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
