+++ core-rules/src/test/java/org/hisp/dhis/android/rules/RuleEventTests.java
@@ -1,5 +1,7 @@
 package org.hisp.dhis.android.rules;
 
+import org.hisp.dhis.android.rules.models.RuleDataValue;
+import org.hisp.dhis.android.rules.models.RuleEvent;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
